fix(desktop): align spend alerts with pricing estimates
