+++ packages/shared/src/index.ts
@@ -50,4 +50,5 @@ export * from "./thread-terminal";
 export * from "./thread-titles";
 export * from "./tool-activity-burst";
 export * from "./token-usage-pricing";
+export * from "./thread-pricing-projection";
 export * from "./worktree-paths";
